backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
docling==2.31.0
httpx>=0.27.0
pydantic==2.9.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
paddleocr>=2.7.0
